move add_exclude_from_nav_index migration step to betas
