+++ plone/app/upgrade/v52/profiles.zcml
@@ -13,10 +13,10 @@
       />
 
   <genericsetup:registerProfile
-      name="to52alpha2"
-      title="Upgrade profile for Plone 5.1.x to Plone 5.2alpha2"
+      name="to52beta1"
+      title="Upgrade profile for Plone 5.2alpha2 to Plone 5.2beta1"
       description=""
-      directory="profiles/to_alpha2"
+      directory="profiles/to_beta1"
       for="Products.CMFPlone.interfaces.IMigratingPloneSiteRoot"
       provides="Products.GenericSetup.interfaces.EXTENSION"
       />
