+++ plone/app/upgrade/v52/profiles/to_beta1/registry.xml
@@ -4,4 +4,5 @@
   <records interface="Products.CMFPlone.interfaces.controlpanel.INavigationSchema" prefix="plone" purge="False">
     <value key="navigation_depth">1</value>
   </records>
+
 </registry>
